requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==3.0.1
django-taggit==6.1.0
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
